Updates application for Android 11

+++ app/src/main/kotlin/com/kelsos/mbrc/annotations/Connection.kt
@@ -1,6 +1,6 @@
 package com.kelsos.mbrc.annotations
 
-import android.support.annotation.IntDef
+import androidx.annotation.IntDef
 
 object Connection {
   const val OFF = 0
@@ -8,6 +8,6 @@ object Connection {
   const val ACTIVE = 2
 
   @Retention(AnnotationRetention.SOURCE)
-  @IntDef(OFF.toLong(), ON.toLong(), ACTIVE.toLong())
+  @IntDef(OFF, ON, ACTIVE)
   annotation class Status
 }//no instance
