+++ app/src/main/kotlin/com/kelsos/mbrc/annotations/Search.kt
@@ -1,6 +1,6 @@
 package com.kelsos.mbrc.annotations
 
-import android.support.annotation.IntDef
+import androidx.annotation.IntDef
 
 object Search {
   const val SECTION_GENRE = 0
@@ -8,7 +8,7 @@ object Search {
   const val SECTION_ALBUM = 2
   const val SECTION_TRACK = 3
 
-  @IntDef(SECTION_GENRE.toLong(), SECTION_ARTIST.toLong(), SECTION_ALBUM.toLong(), SECTION_TRACK.toLong())
+  @IntDef(SECTION_GENRE, SECTION_ARTIST, SECTION_ALBUM, SECTION_TRACK)
   @Retention(AnnotationRetention.SOURCE)
   annotation class Section
 }
