+++ app/src/main/kotlin/com/kelsos/mbrc/annotations/SocketAction.kt
@@ -1,6 +1,6 @@
 package com.kelsos.mbrc.annotations
 
-import android.support.annotation.IntDef
+import androidx.annotation.IntDef
 
 object SocketAction {
   const val RESET = 1
@@ -10,6 +10,6 @@ object SocketAction {
   const val STOP = 5
 
   @Retention(AnnotationRetention.SOURCE)
-  @IntDef(RESET.toLong(), START.toLong(), RETRY.toLong(), TERMINATE.toLong(), STOP.toLong())
+  @IntDef(RESET, START, RETRY, TERMINATE, STOP)
   annotation class Action
 }
